backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.19
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.7.0
email-validator==2.2.0
numpy==2.2.0
opencv-python==4.10.0.84
Pillow==11.0.0
torch==2.5.1
torchvision==0.20.1
facenet-pytorch==2.6.0
matplotlib==3.9.3
scikit-learn==1.6.0
python-dateutil==2.9.0.post0